requirements.txt
Django==5.2.5
asgiref==3.9.1
djangorestframework==3.16.1
sqlparse==0.5.3
